CanvasSwapChain supports full range of present options
Added:

1) Present overload that takes a sync interval
2) Rotation property
3) TransformMatrix property
4) SourceSize property

+++ tests/ExampleGallery/Shared/DpiExample.xaml.cs
@@ -123,7 +123,7 @@ namespace ExampleGallery
             imageSource = new CanvasImageSource(sender, controlSize, controlSize);
             imageControl.Source = imageSource;
 
-            swapChain = new CanvasSwapChain(sender, (int)controlSize, (int)controlSize);
+            swapChain = new CanvasSwapChain(sender, controlSize, controlSize);
             swapChainPanel.SwapChain = swapChain;
         }
